python/pyproject.toml
[project]
name = "promptlab"
version = "0.1.0"
description = "GEPA-based prompt optimization for LLM tasks"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gepa>=0.0.27",
    "pydantic>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "numpy>=1.24",
    "tomli>=2.0",
    "tomli-w>=1.0",
    "datasets>=2.14",
    "python-dotenv>=1.2.1",
    "openai>=2.17.0",
    "litellm>=1.81.9",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.25"]
ollama = ["ollama>=0.2"]
embeddings = ["sentence-transformers>=2.2"]
dev = ["pytest>=8.0", "ruff>=0.3", "pyright>=1.1"]
all = [
    "promptlab[openai,anthropic,ollama,embeddings,dev]",
]

[project.scripts]
promptlab = "promptlab.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "strict"
